update(components/proxy/plugins): move existing plugins to their own modules
