+++ components/proxy/plugins/corsorigin/go.mod
@@ -1,0 +1,8 @@
+module github.com/gitpod-io/gitpod/proxy/plugins/corsorigin
+
+go 1.16
+
+require (
+	github.com/caddyserver/caddy/v2 v2.4.3
+	github.com/rs/cors v1.8.0
+)
