--- components/proxy/plugins/go.mod
@@ -1,8 +1,0 @@
-module github.com/gitpod-io/gitpod/proxy/plugins
-
-go 1.16
-
-require (
-	github.com/caddyserver/caddy/v2 v2.4.1
-	github.com/rs/cors v1.7.0
-)
