+++ components/proxy/plugins/secwebsocketkey/go.mod
@@ -1,0 +1,5 @@
+module github.com/gitpod-io/gitpod/proxy/plugins/secwebsocketkey
+
+go 1.16
+
+require github.com/caddyserver/caddy/v2 v2.4.3
