+++ components/proxy/plugins/secwebsocketkey/sec-websocket-key.go
@@ -2,7 +2,7 @@
 // Licensed under the GNU Affero General Public License (AGPL).
 // See License-AGPL.txt in the project root for license information.
 
-package plugins
+package secwebsocketkey
 
 import (
 	"crypto/rand"
