+++ components/proxy/plugins/workspacedownload/go.mod
@@ -1,0 +1,5 @@
+module github.com/gitpod-io/gitpod/proxy/plugins/workspacedownload
+
+go 1.16
+
+require github.com/caddyserver/caddy/v2 v2.4.3
