+++ components/proxy/plugins/workspacedownload/workspace_download.go
@@ -2,7 +2,7 @@
 // Licensed under the GNU Affero General Public License (AGPL).
 // See License-AGPL.txt in the project root for license information.
 
-package plugins
+package workspacedownload
 
 import (
 	"fmt"
